Resolve-AI/pyproject.toml
[project]
name = "resolve-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
